voizable-backend/ai-service/requirements.txt
# Web API Framework
fastapi
uvicorn[standard]
python-multipart

# AI and Transcription
openai-whisper
torch

# Dependencies (can be kept if auto-added)
Jinja2~=3.1.3
typing_extensions~=4.9.0